Reorganize view tests in separate files
The test_views_api was 2500+ lines long
and became quite a jungle to get through..
